requirements.txt
flask
beautifulsoup4
pandas
langchain
langchain-community
langchain-huggingface
langchain-ollama
langchain-core
sentence-transformers>=3.0.0
tabulate
json5
scikit-learn
tiktoken
openpyxl
numpy
rapidfuzz
nltk
requests
re  # stdlib, but often mentioned for clarity
textwrap  # stdlib
python-dotenv

